requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
PyJWT==2.7.0
pymysql==1.1.0
Werkzeug==2.3.7
gunicorn==21.2.0
cloud-sql-python-connector==0.10.1
